requirements.txt
fpdf==2.7.6
openai==1.12.0
python-docx==1.1.0
python-dotenv==1.0.1
streamlit==1.32.0 